document-qa/requirements.txt
streamlit>=1.24.0
Pillow>=9.5.0
pdf2image>=1.16.3
pytesseract>=0.3.10
numpy>=1.24.3
torch>=2.0.0
transformers>=4.30.2
datasets>=2.13.0
opencv-python>=4.7.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain>=0.0.200
google-cloud-aiplatform>=1.35.0
google-generativeai>=0.3.1
